fix(player): make mini player window draggable
