+++ Sources/Kaset/Views/YouTube/YouTubeVideoWindowController.swift
@@ -628,55 +628,3 @@ private struct YouTubeVideoWindowContent: View {
         )
     }
 }
-
-// MARK: - WindowDragHandle
-
-/// Transparent native strip that lets the user move the floating window by
-/// dragging along the top. The hosted WebView reports
-/// `mouseDownCanMoveWindow == false` and consumes `mouseDown`, defeating the
-/// window's `isMovableByWindowBackground` everywhere it covers; this strip sits
-/// above the WebView and drives the move explicitly through
-/// `NSWindow.performDrag(with:)`. Scoped to the floating window only — the
-/// shared `YouTubeWatchSurfaceView` is untouched.
-private struct WindowDragHandle: NSViewRepresentable {
-    func makeNSView(context _: Context) -> NSView {
-        WindowDragNSView()
-    }
-
-    func updateNSView(_: NSView, context _: Context) {}
-}
-
-// MARK: - WindowDragNSView
-
-/// Backing view for `WindowDragHandle`.
-private final class WindowDragNSView: NSView {
-    /// Take `mouseDown` ourselves instead of letting AppKit's background-drag
-    /// heuristics intercept it, so the move is driven deterministically.
-    override var mouseDownCanMoveWindow: Bool {
-        false
-    }
-
-    /// Drag even when the floating window is not key — it is ordered front
-    /// without stealing focus, so a first click must move it, not just activate.
-    override func acceptsFirstMouse(for _: NSEvent?) -> Bool {
-        true
-    }
-
-    override func mouseDown(with event: NSEvent) {
-        // Preserve the standard titlebar gesture: a double-click performs the
-        // user's configured "double-click a window's title bar to" action
-        // (Zoom / Minimize / None); a single click starts the window drag.
-        if event.clickCount == 2 {
-            switch UserDefaults.standard.string(forKey: "AppleActionOnDoubleClick") {
-            case "Minimize":
-                self.window?.miniaturize(nil)
-            case "None":
-                break
-            default: // "Maximize" (zoom) is the macOS default.
-                self.window?.performZoom(nil)
-            }
-        } else {
-            self.window?.performDrag(with: event)
-        }
-    }
-}
